replace NULL with nullptr

+++ simulation/g4simulation/g4jets/FastJetAlgo.h
@@ -11,7 +11,7 @@ class FastJetAlgo : public JetAlgo {
   
 public:
 
-  FastJetAlgo(Jet::ALGO algo, float par, float verbosity = 0);
+  FastJetAlgo(Jet::ALGO algo, float par, int verbosity = 0);
   virtual ~FastJetAlgo() {}
 
   void      identify(std::ostream& os = std::cout);
